feat: structure the src folder to contain only vuse core code and remove the demo project
