--- src/pug/index.pug
@@ -1,4 +1,0 @@
-extends layout.pug
-
-append content
-  b-builder(@saved="onSave")
