--- src/pug/layout.pug
@@ -1,10 +1,0 @@
-html(lang="en")
-  head
-    meta(charset="UTF-8")
-    meta(name="viewport", content="width=device-width, initial-scale=1.0")
-    meta(http-equiv="X-UA-Compatible", content="ie=edge")
-    title Document
-  body
-    include mixin/icon
-    #app
-      block content
